Reorganize tests by module
